// src/core/api/gql/lottery/get-lotteries.gql.generated.ts
export type Maybe<T> = T | null;
export type InputMaybe<T> = Maybe<T>;
export type Exact<T extends { [key: string]: unknown }> = {
  [K in keyof T]: T[K];
};
// ...
export type Scalars = {
  ID: { input: string; output: string };
  String: { input: string; output: string };
  Boolean: { input: boolean; output: boolean };
  Int: { input: number; output: number };
  Float: { input: number; output: number };
  Decimal: { input: BigNumber; output: BigNumber };
};

export type AccountType = {
  __typename?: "AccountType";
  address: Scalars["String"]["output"];
  avatar?: Maybe<Scalars["String"]["output"]>;
  createdAt: Scalars["String"]["output"];
  name?: Maybe<Scalars["String"]["output"]>;
  twitter?: Maybe<Scalars["String"]["output"]>;
  updatedAt: Scalars["String"]["output"];
};

export type AssetType = {
  __typename?: "AssetType";
  address: Scalars["String"]["output"];
  createdAt: Scalars["String"]["output"];
  decimals?: Maybe<Scalars["Int"]["output"]>;
  deprecated: Scalars["Boolean"]["output"];
  id: Scalars["String"]["output"];
  name: Scalars["String"]["output"];
  symbol: Scalars["String"]["output"];
  type: Scalars["String"]["output"];
};

export type AuthType = {
  __typename?: "AuthType";
  accessToken: Scalars["String"]["output"];
  account?: Maybe<AccountType>;
  address: Scalars["String"]["output"];
};

export enum DrawStatus {
  Cancelled = "CANCELLED",
  Completed = "COMPLETED",
  Pending = "PENDING",
}

export type DrawType = {
  __typename?: "DrawType";
  drawDate?: Maybe<Scalars["String"]["output"]>;
  id: Scalars["String"]["output"];
  lotteryId: Scalars["String"]["output"];
  status: DrawStatus;
  transactionHash?: Maybe<Scalars["String"]["output"]>;
  updatedAt: Scalars["String"]["output"];
  winner?: Maybe<AccountType>;
};

export type LinkTwitterInput = {
  code: Scalars["String"]["input"];
  state: Scalars["String"]["input"];
};

export type LoginSignatureInput = {
  /** Signature for provided timestamp */
  signature: Scalars["String"]["input"];
  /** Timestamp used for the signature */
  timestamp: Scalars["String"]["input"];
};

export type LotteryFilterInput = {
  featured?: InputMaybe<Scalars["Boolean"]["input"]>;
  status?: InputMaybe<LotteryStatus>;
  uid?: InputMaybe<Scalars["String"]["input"]>;
};

export enum LotteryStatus {
  Cancelled = "CANCELLED",
  Completed = "COMPLETED",
  Ongoing = "ONGOING",
  Scheduled = "SCHEDULED",
}

export type LotteryType = {
  __typename?: "LotteryType";
  createdAt: Scalars["String"]["output"];
  draw?: Maybe<DrawType>;
  endDate: Scalars["String"]["output"];
  feeTicketAmount: Scalars["Decimal"]["output"];
  id: Scalars["String"]["output"];
  maxTickets: Scalars["Int"]["output"];
  name: Scalars["String"]["output"];
  prize: PrizeType;
  startDate: Scalars["String"]["output"];
  status: LotteryStatus;
  ticketAsset?: Maybe<AssetType>;
  ticketPrice: Scalars["Decimal"]["output"];
  tickets: Array<TicketType>;
  uid: Scalars["String"]["output"];
  updatedAt: Scalars["String"]["output"];
};
// ...
export enum PrizeStatus {
  Active = "ACTIVE",
  Distributed = "DISTRIBUTED",
  Refunded = "REFUNDED",
}

export type PrizeType = {
  __typename?: "PrizeType";
  createdAt: Scalars["String"]["output"];
  id: Scalars["String"]["output"];
  lotteryId: Scalars["String"]["output"];
  prizeAsset: AssetType;
  status: PrizeStatus;
  totalPrizePool: Scalars["Decimal"]["output"];
  updatedAt: Scalars["String"]["output"];
};

export type Query = {
  __typename?: "Query";
  /** Get account graph by address */
  account: AccountType;
  /** Get list of all supported assets */
  assets: Array<AssetType>;
  /** Fetches the last 10 tickets sold */
  getLastTickets: Array<TicketType>;
  getUserTickets: Array<TicketType>;
  /** Get list of all supported assets */
  lotteries: Array<LotteryType>;
  lottery: LotteryType;
  twitterAuthUrl: Scalars["String"]["output"];
  verify?: Maybe<DrawType>;
};
// ...
export type TicketType = {
  __typename?: "TicketType";
  account: AccountType;
  asset: AssetType;
  createdAt: Scalars["String"]["output"];
  id: Scalars["String"]["output"];
  /** Represent the price of each ticket */
  individualTicketPrice: Scalars["Decimal"]["output"];
  lotteryId: Scalars["String"]["output"];
  lotteryUid: Scalars["String"]["output"];
  /** Represent the number of tickets the user has bought on this transaction */
  nTickets: Scalars["Int"]["output"];
  purchasedAt: Scalars["String"]["output"];
  /** Represent the total price of all the tickets bought */
  totalPrice: Scalars["Decimal"]["output"];
  transactionHash: Scalars["String"]["output"];
  updatedAt: Scalars["String"]["output"];
};

export type UpdateAccountInput = {
  bio?: InputMaybe<Scalars["String"]["input"]>;
  name?: InputMaybe<Scalars["String"]["input"]>;
  slug?: InputMaybe<Scalars["String"]["input"]>;
  twitter?: InputMaybe<Scalars["String"]["input"]>;
};

export type GetLotteriesQueryVariables = Types.Exact<{
  filter: Types.LotteryFilterInput;
}>;

export type GetLotteriesQuery = {
  __typename?: "Query";
  lotteries: Array<{
    __typename?: "LotteryType";
    createdAt: string;
    endDate: string;
    feeTicketAmount: BigNumber;
    id: string;
    maxTickets: number;
    name: string;
    startDate: string;
    status: Types.LotteryStatus;
    ticketPrice: BigNumber;
    uid: string;
    updatedAt: string;
    tickets: Array<{
      __typename?: "TicketType";
      createdAt: string;
      updatedAt: string;
      lotteryId: string;
      lotteryUid: string;
      id: string;
      transactionHash: string;
      purchasedAt: string;
      nTickets: number;
      totalPrice: BigNumber;
      individualTicketPrice: BigNumber;
      asset: {
        __typename?: "AssetType";
        address: string;
        createdAt: string;
        decimals?: number | null;
        deprecated: boolean;
        id: string;
        name: string;
        symbol: string;
        type: string;
      };
      account: {
        __typename?: "AccountType";
        twitter?: string | null;
        address: string;
        name?: string | null;
        avatar?: string | null;
        createdAt: string;
        updatedAt: string;
      };
    }>;
    draw?: {
      __typename?: "DrawType";
      drawDate?: string | null;
      id: string;
      lotteryId: string;
      status: Types.DrawStatus;
      updatedAt: string;
      transactionHash?: string | null;
      winner?: {
        __typename?: "AccountType";
        twitter?: string | null;
        address: string;
        name?: string | null;
        avatar?: string | null;
        createdAt: string;
        updatedAt: string;
      } | null;
    } | null;
    prize: {
      __typename?: "PrizeType";
      createdAt: string;
      updatedAt: string;
      status: Types.PrizeStatus;
      lotteryId: string;
      id: string;
      totalPrizePool: BigNumber;
      prizeAsset: {
        __typename?: "AssetType";
        address: string;
        createdAt: string;
        decimals?: number | null;
        deprecated: boolean;
        id: string;
        name: string;
        symbol: string;
        type: string;
      };
    };
    ticketAsset?: {
      __typename?: "AssetType";
      address: string;
      createdAt: string;
      decimals?: number | null;
      deprecated: boolean;
      id: string;
      name: string;
      symbol: string;
      type: string;
    } | null;
  }>;
};